split server and wrapper into different files
